Add support for local dependencies
When Nimble unit tests related to package dependencies are being
written, it is convenient to be able to install dependencies from the
local machine on which the tests are being executed, instead of relying
on packages downloaded over network connection. This has also an
advantage of faster tests execution.

This feature is implemented for being used in writing a test case for
the problem with multiple downloads and installs of a single dependency.

Related to: #678