Fix getting download dir name for develop
When getting the download directory name for a package that we want to
put in develop mode, a trailing slash in the URL or the subdir leads to
getting an empty string. To fix the problem we remove the trailing slash
from them at the start of the `getDevelopDownloadDir` procedure.