Omit cloning if the package is on local
When developing local dependent packages, the unmodified implementation should use

```console
nimble develop file:///absolute/path/to/package
```

but this is not suitable for development since it will be cloned into a subdirectory.

If the package is a local package, it is easier to handle it as is without cloning.

With this change,

```console
nimble develop file:///absolute/path/to/package --withDependencies
```

would change nimble.develop as follows:

```diff
"dependencies": [
  ...
+ "absolute/path/to/package"
]
```