a more elegant way to retrieve the nim right version
Note we may want to have the solvedPackages inside Options but a refactor would be needed for this and first we should remove the support for the `legacy` solver. Everything is build around it ( i.e. the `legacy` solver doesnt have enough information to split the download stage from the build one)