zero to hero: nimble is able to function without nim in the path (#1269)
* when in `solver:sat` sets the right `nim` version

* a more elegant way to retrieve the nim right version

Note we may want to have the solvedPackages inside Options but a refactor would be needed for this and first we should remove the support for the `legacy` solver. Everything is build around it ( i.e. the `legacy` solver doesnt have enough information to split the download stage from the build one)

* remove comment

* Should use the system nim when `--useSystemNim` flag is on fixes #1251.

fixes #1251

* zero to hero: `nimble` is able to function without `nim` in the path

* fixes test comp