Introduce compat/ to work around nim bugs
Depending on the Nim version used to compile nimble, we'll encounter a
different set of bugs - `compat/` is an easy way to deal with them and
not have to spread compatiblity code all over the codebase (like the
ugly layers of exception handlers).

Often, Exception effects come from forward declarations and a few
strategically placed explicit raises list can do wonders to limit  the
spread of workarounds.